Pull together the two builders since they were now sharing a lot of resources
